backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
dataclasses-json==0.6.7
distro==1.9.0
Django==4.2.13
django-cors-headers==4.4.0
djangorestframework==3.15.1
dnspython==2.6.1
exceptiongroup==1.2.1
faiss-cpu==1.8.0.post1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.11
langchain-openai==0.1.14
langchain-text-splitters==0.2.2
langsmith==0.1.83
marshmallow==3.21.3
mongoengine==0.28.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.24.4
openai==1.35.10
orjson==3.10.6
packaging==24.1
pydantic==2.8.2
pydantic-core==2.20.1
PyJWT==2.8.0
pymongo==4.7.3
pypdf2==3.0.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlparse==0.5.0
tenacity==8.4.2
tiktoken==0.7.0
tqdm==4.66.4
typing-extensions==4.12.1
typing-inspect==0.9.0
urllib3==2.2.2
yarl==1.9.4
